Started reworking TrendWidget
